MOBI-489 Showcase Updates - fix NPE on shutdown in Tomcat6
git-svn-id: http://anonsvn.icesoft.org/repo/icemobile/trunk/icemobile@33048 1de00972-ac04-4911-a46b-06793c8d78d9

+++ samples/mobileshowcase/src/main/java/org/icemobile/samples/mobileshowcase/view/examples/device/microphone/MicrophoneBean.java
@@ -135,9 +135,11 @@ public class MicrophoneBean extends ExampleImpl<MicrophoneBean> implements
 
     @PreDestroy
     public void disposeResources(){
-        boolean success = audioFile.delete();
-        if (!success && logger.isLoggable(Level.FINE)){
-            logger.fine("Could not dispose of media file" + audioFile.getAbsolutePath());
+        if( audioFile != null ){
+            boolean success = audioFile.delete();
+            if (!success && logger.isLoggable(Level.FINE)){
+                logger.fine("Could not dispose of media file" + audioFile.getAbsolutePath());
+            }
         }
     }
